OrderSerializationSchema implementation for test serializer interface
